Add crop x/y/w/h params to PictureReady
For nearly all video codecs, buffers that are potentially reference
frames require some amount of paddng around the edges.  Partially
to handle decoder alignment requirements, but primarily to deal with
macroblocks that move off the edge of the visible picture, and then
back into view in a future frame.  Some of the ffmpeg/libav sw
codecs handle this by storing the non-visible pixels in a seperate
buffer (at the cost of some runtime overhead).  But many hardware
decoders cannot support this.  One solution is to introce an extra
copy of the output buffer (sometimes combined with conversion to
RGB).  Although if the GPU can handle sampling of YUV textures, this
extra copy is unnecessary and wastes power.

To avoid introducing an extra copy on the output buffer of the
decoder, it is preferable to pass the x/y offset and cropped width
and height of the output buffer of the decoder through to the
renderer, and handle the cropping during rendering.

Change-Id: I9af08a6c1a7db62101d9abb58aed606b3ee15047